refactor: move files to the kubernetes sub folder
