+++ packages/main/src/plugin/kubernetes/kubernetes-port-forward-validation.ts
@@ -15,7 +15,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  ***********************************************************************/
-import type { ForwardConfig } from '/@/plugin/kubernetes-port-forward-model.js';
+import type { ForwardConfig } from '/@/plugin/kubernetes/kubernetes-port-forward-model.js';
 
 /**
  * Type representing a function that checks the availability of a port.
